adding new module pokemon and refactoring the code for new module
