+++ src/app/pokemon/border-card.directive.ts
