+++ src/app/pokemon/detail-pokemon/detail-pokemon.component.html
