+++ src/app/pokemon/list-pokemon/list-pokemon.component.html
