+++ src/app/pokemon/pokemon-type-color.pipe.ts
